switching email_id to bigint

+++ database/migrations/2016_06_29_161639_create_attribution_activity_schedules_table.php
@@ -16,7 +16,7 @@ class CreateAttributionActivitySchedulesTable extends Migration
     public function up()
     {
         Schema::create('attribution_activity_schedules', function (Blueprint $table) {
-            $table->integer( 'email_id' )->unsigned();
+            $table->bigInteger( 'email_id' )->unsigned();
             $table->date( 'inactive_date' );
             $table->timestamps();
